safari-600.7-branch/LayoutTests/fullscreen/full-screen-restrictions-expected.txt

This tests the restrictions to entering full screen mode laid out in section 4.1 of the W3C Full Screen API
"The context object is not in a document."
EVENT(webkitfullscreenerror)
"The context object's node document, or an ancestor browsing context's document does not have the fullscreen enabled flag set."
EVENT(webkitfullscreenerror)
"The context object's node document fullscreen element stack is not empty and its top element is not an ancestor of the context object."
EVENT(webkitfullscreenchange)
EVENT(webkitfullscreenerror)
EVENT(webkitfullscreenchange)
"A descendant browsing context's document has a non-empty fullscreen element stack."
EVENT(webkitfullscreenchange)
EVENT(webkitfullscreenerror)
EVENT(webkitfullscreenchange)
"This algorithm is not allowed to show a pop-up."
EVENT(webkitfullscreenerror)
END OF TEST
